+++ src/vs/editor/browser/controller/editContext/native/nativeEditContextUtils.ts
@@ -3,7 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { EditContextEventHandlersEventMap } from './editContext.js';
 import { addDisposableListener } from '../../../../../base/browser/dom.js';
 import { IDisposable, Disposable } from '../../../../../base/common/lifecycle.js';
